+++ inc/core/ghost.h
@@ -49,6 +49,7 @@ public:
   void eaten(std::shared_ptr<Map> map);
 
   void move(std::shared_ptr<Map> map);
+  bool eat_entity(std::shared_ptr<Map> map);
 };
 
 #endif // __inc_core_ghost_H__