+++ inc/state.h
@@ -4,7 +4,7 @@
 
 #include <iostream>
 
-enum class GameState { MENU, GAME, PAUSE, GAME_OVER, EXIT };
+enum class GameState { MENU, WAITING, GAME, PAUSE, GAME_OVER, EXIT };
 enum class PcmnState { ALIVE, DEAD, POWERED };
 enum class GhstState { CHASE, SCATTER, FRIGHTENED, EATEN };