+++ src/core/ghost.cpp
@@ -23,9 +23,8 @@ void Ghost::show(std::shared_ptr<Renderer> renderer) {
   switch (state) {
 
   case GhstState::FRIGHTENED:
-    // todo: ending animation (pass `true` to get_sprite_ghost_weak)
     asset = renderer->get_assets()->get_sprite_ghost_weak(
-      renderer->get_fps_anim_count());
+      renderer->get_fps_anim_count(), p_timer->step_passed(7));
     break;
 
   case GhstState::EATEN:
@@ -65,8 +64,11 @@ bool Ghost::can_go(std::shared_ptr<Map> map, const Direction &dir) const {
   std::tie(i, j) = get_ij(map->get_size());
   if (i == -1 || j == -1) { return false; }
 
-  return map->can_go(j, i, dir); // why j, i and not i, j ?
-                                 // well ...
+  if (map->is_home(i, j) || state == GhstState::EATEN) {
+    return map->can_go(j, i, dir, true);
+  } else {
+    return map->can_go(j, i, dir);
+  }
 }
 
 void Ghost::blinky_chase(std::shared_ptr<Map> map,
@@ -177,7 +179,20 @@ void Ghost::scatter(std::shared_ptr<Map> map) {
   m_reg_direction = map->stupid(ghost_tile, target, get_direction());
 }
 
-void Ghost::frightened(std::shared_ptr<Map> map) { (void)map; }
+void Ghost::frightened(std::shared_ptr<Map> map) {
+  auto [i, j] = get_ij(map->get_size());
+  Node ghost_tile = {j, i};
+
+  std::vector<Direction> possible_directions;
+  std::vector<Node> possible_nodes = map->get_neighbors(ghost_tile);
+  for (auto node : possible_nodes) {
+    possible_directions.push_back(get_direction_from_nodes(ghost_tile, node));
+  }
+
+  std::uniform_int_distribution<> dist(0, possible_directions.size() - 1);
+  int random_index = dist(gen);
+  m_reg_direction = possible_directions[random_index];
+}
 
 void Ghost::eaten(std::shared_ptr<Map> map) { (void)map; }
 
@@ -185,26 +200,59 @@ void Ghost::update(std::shared_ptr<Map> map, std::tuple<int, int> pacman_pos,
                    Direction pacman_dir) {
 
   teleport(map);
+  p_timer = map->get_power_timer();
 
   // start or init the timer the first time
-  if (!m_timer.is_running()) { m_timer.start_timer(7); }
-
-  // change state
-  if (m_timer.is_expired()) {
-    switch (state) {
-    case GhstState::SCATTER:
-      state = GhstState::CHASE;
-      m_timer.reset_timer();
-      m_timer.start_timer(20);
-      break;
-    case GhstState::CHASE:
-      state = GhstState::SCATTER;
-      m_timer.reset_timer();
+  if (!m_timer.is_running()) {
+    if (is_at_home) {
+      m_timer.start_timer(dis(gen));
+    } else {
       m_timer.start_timer(7);
-      break;
-    default: break; // todo: will become fmt::unreachable
     }
   }
+
+  if (!is_at_home && state != GhstState::FRIGHTENED && map->pcmn_powered()) {
+    state = GhstState::FRIGHTENED;
+    m_timer.reset_timer(); // reset to use global timer
+  }
+
+  // change state
+  if ((m_timer.is_running() && m_timer.is_expired()) || p_timer->is_expired()) {
+    if (is_at_home) {
+      is_at_home = false;
+      m_timer.reset_timer();
+      m_timer.start_timer(7);
+      state = GhstState::SCATTER; // not needed ?
+    } else {
+      switch (state) {
+      case GhstState::SCATTER:
+        state = GhstState::CHASE;
+        m_timer.reset_timer();
+        m_timer.start_timer(20);
+        break;
+      case GhstState::CHASE:
+        state = GhstState::SCATTER;
+        m_timer.reset_timer();
+        m_timer.start_timer(7);
+        break;
+      case GhstState::FRIGHTENED:
+        state = GhstState::SCATTER;
+        m_timer.reset_timer(); // reset for ghosts that just went out of the
+                               // house
+        m_timer.start_timer(7);
+        break;
+      default: fmt::unreachable("Invalid ghost state when timer expired");
+      }
+    }
+  }
+  // go out of frightened state
+  if (state == GhstState::FRIGHTENED && !p_timer->is_running()) {
+    state = GhstState::SCATTER;
+    m_timer.reset_timer();
+    m_timer.start_timer(7);
+  }
+
+  auto [i, j] = get_ij(map->get_size());
 
   if (is_at_home) {
 
@@ -223,6 +271,11 @@ void Ghost::update(std::shared_ptr<Map> map, std::tuple<int, int> pacman_pos,
     case GhstState::FRIGHTENED: frightened(map); break;
     case GhstState::EATEN: eaten(map); break;
     }
+    // override the registered direction if the ghost is still at home
+    if (map->is_home(i, j) /* here it is still i, j */) {
+      m_reg_direction = map->astar({j, i}, map->get_door_node());
+    }
+
     // if the registered direction is not the opposite of the current direction
     // and the ghost can go in that direction
     if (m_reg_direction != Direction::NONE && can_change_direction(map) &&
@@ -251,6 +304,16 @@ void Ghost::move(std::shared_ptr<Map> map) {
   default: break;
   }
 
-  // save the position of blinky because we need it for inky
-  if (type == GhostType::BLINKY) { map->set_blinky_pos(m_cx, m_cy); }
+  switch (type) {
+  case GhostType::BLINKY: map->set_blinky_pos(m_cx, m_cy); break;
+  case GhostType::PINKY: map->set_pinky_pos(m_cx, m_cy); break;
+  case GhostType::INKY: map->set_inky_pos(m_cx, m_cy); break;
+  case GhostType::CLYDE: map->set_clyde_pos(m_cx, m_cy); break;
+  default: fmt::unreachable("Invalid ghost type");
+  }
+}
+
+bool Ghost::eat_entity(std::shared_ptr<Map> map) {
+  (void)map;
+  return false;
 }