+++ src/core/pcmn.cpp
@@ -9,7 +9,7 @@ Pacman::Pacman(const double cx, const double cy,
                const struct PacmanConfig &config)
   : Entity{cx, cy, 0, 0}, m_points_per_dot{config.m_points_per_dot},
     m_points_per_power_dot{config.m_points_per_power_dot},
-    m_points_per_ghost{config.m_points_per_ghost} {}
+    m_points_per_ghost{config.m_points_per_ghost}, state{PcmnState::ALIVE} {}
 
 void Pacman::show(std::shared_ptr<Renderer> renderer) {
   static const double custom_scale = 0.5;
@@ -35,26 +35,48 @@ bool Pacman::can_go(std::shared_ptr<Map> map, const Direction &dir) const {
 }
 
 bool Pacman::ate_food(std::shared_ptr<Map> map) {
-  int i = -1, j = -1;
-  std::tie(i, j) = get_ij(map->get_size());
-  if (i == -1 || j == -1) { return false; }
-
+  auto [i, j] = get_ij(map->get_size());
   return map->ate_food(j, i); // why j, i and not i, j ? well ...
 }
-
 void Pacman::eat_food(std::shared_ptr<Map> map) {
-  int i = -1, j = -1;
-  std::tie(i, j) = get_ij(map->get_size());
-  if (i == -1 || j == -1) { return; }
-
+  auto [i, j] = get_ij(map->get_size());
   map->eat_food(j, i); // those damn vectors
 
   m_score += m_points_per_dot;
 }
 
+bool Pacman::ate_big_food(std::shared_ptr<Map> map) {
+  auto [i, j] = get_ij(map->get_size());
+  return map->ate_big_food(j, i);
+}
+void Pacman::eat_big_food(std::shared_ptr<Map> map) {
+  auto [i, j] = get_ij(map->get_size());
+  map->eat_big_food(j, i);
+
+  m_score += m_points_per_power_dot;
+  state = PcmnState::POWERED;
+}
+
 void Pacman::update(std::shared_ptr<Map> map) {
 
   teleport(map);
+  p_timer = map->get_power_timer();
+
+  if (state == PcmnState::POWERED && !p_timer->is_running()) {
+    p_timer->start_timer(10);
+  }
+
+  // here we test if the timer is running before because, as opposed to ghosts
+  // pacman behavior is not always being timed
+  if (p_timer->is_running() && p_timer->is_expired()) {
+    switch (state) {
+    case PcmnState::POWERED:
+      state = PcmnState::ALIVE;
+      p_timer->reset_timer();
+      break;
+    default: break;
+    }
+  }
 
   // first check if we can still go in the current direction
   if (can_change_direction(map) && !can_go(map, m_direction)) {
@@ -79,7 +101,12 @@ void Pacman::update(std::shared_ptr<Map> map) {
       can_go(map, opposite(m_direction))) {
     m_direction = opposite(m_direction);
   }
+  move(map);
+  if (ate_food(map)) { eat_food(map); }
+  if (ate_big_food(map)) { eat_big_food(map); }
+}
 
+void Pacman::move(std::shared_ptr<Map> map) {
   switch (m_direction) {
   case Direction::UP: m_cy -= m_speed; break;
   case Direction::DOWN: m_cy += m_speed; break;
@@ -87,5 +114,18 @@ void Pacman::update(std::shared_ptr<Map> map) {
   case Direction::RIGHT: m_cx += m_speed; break;
   default: break;
   }
-  if (ate_food(map)) { eat_food(map); }
+  map->set_pacman_pos(m_cx, m_cy);
+  map->pcmn_powered(state == PcmnState::POWERED);
+}
+
+bool Pacman::eat_entity(std::shared_ptr<Map> map) {
+  (void)map;
+  switch (state) {
+  case PcmnState::ALIVE:
+    m_lives--;
+    state = PcmnState::DEAD;
+    return m_lives > 0;
+  case PcmnState::POWERED: m_score += m_points_per_ghost; return true;
+  default: fmt::unreachable("Pacman::eat_entity : dead pacman cannot eat");
+  }
 }